+++ src/Hacks/Noclip/Noclip.cpp
@@ -34,6 +34,12 @@ void NoclipPlayLayer::destroyPlayer(PlayerObject* p0, GameObject* p1)
     else
     {
         SafeMode::get()->setHackedAttempt();
+
+        if (!m_fields->hasDied)
+        {
+            m_fields->hasDied = true;
+            m_fields->timeDead += CCDirector::get()->getDeltaTime();
+        }
         
         m_fields->isDead = true;
 
@@ -62,5 +68,24 @@ void NoclipPlayLayer::resetLevel()
 
     m_fields->d = 0;
     m_fields->t = 0;
+    m_fields->timeDead = 0;
     m_fields->isDead = false;
+}
+
+float NoclipPlayLayer::getNoclipAccuracy()
+{
+    if (m_gameState.m_currentProgress == 0)
+        return 1;
+
+    return 1 - (m_fields->timeDead / as<float>(m_gameState.m_unkDouble2));
+}
+
+void NoclipBaseGameLayer::update(float dt)
+{
+    GJBaseGameLayer::update(dt);
+
+    if (auto layer = typeinfo_cast<PlayLayer*>(this))
+    {
+        as<NoclipPlayLayer*>(layer)->m_fields->hasDied = false;
+    }
 }